Move class into different package
